refactor: isolate request origin policy
